↺ svn merge -r 1880:1944 https://amethyst.yui.ne.jp/svn-dev/pc98/np2/trunk
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@1946 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ support_gimic/generic/hostdrv.c
@@ -1323,7 +1323,7 @@ static BOOL fhdl_wr(void *vpItem, void *vpArg) {
 	UINT	len;
 
 	p = ((HDRVFILE)vpItem)->path;
-	len = OEMSTRLEN(p);
+	len = (UINT)OEMSTRLEN(p);
 	statflag_write((STFLAGH)vpArg, &len, sizeof(len));
 	if (len) {
 		if (len < MAX_PATH) {
@@ -1337,10 +1337,10 @@ static BOOL fhdl_wr(void *vpItem, void *vpArg) {
 static BOOL flist_wr(void *vpItem, void *vpArg) {
 
 	OEMCHAR	*p;
-	int		len;
+	UINT	len;
 
 	p = ((HDRVLST)vpItem)->realname;
-	len = OEMSTRLEN(p);
+	len = (UINT)OEMSTRLEN(p);
 	if (len < MAX_PATH) {
 		ZeroMemory(p + len, (MAX_PATH - len) * sizeof(OEMCHAR));
 	}
