+++ support_gimic/sdl2/ext/gimic/gimicusb.cpp
@@ -96,7 +96,7 @@ int CGimicUSB::Deinitialize()
  * @param[in] cbInput The maximum number of bytes to be read
  * @return C86CTL_ERR
  */
-int CGimicUSB::Transaction(const void* lpOutput, size_t cbOutput, void* lpInput, size_t cbInput)
+int CGimicUSB::Transaction(const void* lpOutput, int cbOutput, void* lpInput, int cbInput)
 {
 	if (!m_usb.IsOpened())
 	{
@@ -127,7 +127,7 @@ int CGimicUSB::Transaction(const void* lpOutput, size_t cbOutput, void* lpInput,
 
 	if ((lpInput != NULL) && (cbInput > 0))
 	{
-		cbInput = std::min(cbInput, sizeof(sBuffer));
+		cbInput = std::min(cbInput, static_cast<int>(sizeof(sBuffer)));
 		::memcpy(lpInput, sBuffer, cbInput);
 	}
 	return C86CTL_ERR_NONE;
@@ -485,7 +485,7 @@ bool CGimicUSB::Task()
 {
 	/* builds data */
 	UINT8 sData[64];
-	size_t nIndex = 0;
+	int nIndex = 0;
 
 	m_queGuard.Enter();
 	while (m_nQueCount)
